+++ source/libgcl/libgcl.h
@@ -7,6 +7,7 @@
 #define __MGS_LIBGCL_H__
 
 #include <sys/types.h>
+#include <stdint.h>   // for intptr_t
 #include <libgte.h>     // for SVECTOR
 #include <libgpu.h>
 #include "common.h"
@@ -168,7 +169,7 @@ int GCL_Expr( char *data, int *value );
 
 /* parse.c */
 void  GCL_SetArgTop( char *top );
-char *GCL_GetNextValue( char *top, int *type_p, int *value_p );
+char *GCL_GetNextValue( char *top, int *type_p, intptr_t *value_p );
 void *GCL_SetArgStack( GCL_ARGS *args );
 void  GCL_UnsetArgStack( void *stack );
 int   GCL_GetArgs( int argno );
@@ -192,7 +193,7 @@ void  GCL_InitVar( void );
 void  GCL_InitClearVar( void );
 void  GCL_SaveVar( void );
 void  GCL_RestoreVar( void );
-char *GCL_GetVar( char *top, int *type_p, int *value_p );
+char *GCL_GetVar( char *top, int *type_p, intptr_t *value_p );
 char *GCL_SetVar( char *top, int value );
 char *GCL_VarSaveBuffer( char *top );
